Bump to v0.1.0
Pin the `knossos` gem to the matching minor version, as that package's
API is rapidly changing at this point.

As far as practical, we'll try to keep the versions of these two gems in
sync, keeping in mind that our semantic versioning policy permits
"anything goes" in major version 0.

Also, importantly, correctly specify the executable in the .gemspec so
that it will be available on the command line after installing this gem.